requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
python-dotenv>=0.19.0

# Data fetching and API
requests>=2.26.0
yfinance>=0.2.28
python-binance>=1.0.19

# Machine learning and deep learning
joblib>=1.0.1
cachetools>=5.3.1

# Web framework and API
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0

# Database and ORM
tortoise-orm>=0.20.0
aerich>=0.7.2
aiosqlite>=0.19.0

# Authentication and security
PyJWT>=2.4.0
passlib[bcrypt]>=1.7.4
email-validator>=1.1.3

# Utilities
pytz>=2021.1
tqdm>=4.62.2
python-dateutil>=2.8.2
aiohttp>=3.8.0

# Testing and development
pytest>=6.2.5
pytest-asyncio>=0.15.1
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.1

# Added from the code block
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=3.2.0
pydantic>=1.8.2
python-multipart>=0.0.5
alembic>=1.7.1
httpx>=0.18.2
ta>=0.10.1
seaborn>=0.11.2
plotly>=5.1.0

# Additional data sources
alpha-vantage>=2.3.1

# Background task processing (optional)
celery[redis]>=5.2.0
redis>=4.3.0
